clean: createSushiListをコンポーネントで管理
