+++ sushi-front/src/app/api/manageSushiList/route.ts
@@ -6,7 +6,6 @@ import path from 'path';
 const sushiListPath = path.join(process.cwd(), 'src', 'utils', 'sushiList.ts');
 
 export async function GET(req: NextRequest) {
-  // GETリクエストで寿司リストを取得
   try {
     return NextResponse.json(sushiList);
   } catch (error) {
@@ -16,14 +15,12 @@ export async function GET(req: NextRequest) {
 }
 
 export async function PUT(req: NextRequest) {
-  // PUTリクエストで寿司リストを更新
   try {
     const { index, updatedSushi } = await req.json();
 
     if (index >= 0 && index < sushiList.length) {
       sushiList[index] = updatedSushi;
 
-      // ファイルを書き換え
       const updatedContent = `export const sushiList = ${JSON.stringify(sushiList, null, 2)};`;
       fs.writeFileSync(sushiListPath, updatedContent, 'utf8');
 
@@ -38,14 +35,12 @@ export async function PUT(req: NextRequest) {
 }
 
 export async function DELETE(req: NextRequest) {
-  // DELETEリクエストで寿司リストから指定のエントリを削除
   try {
     const { index } = await req.json();
 
     if (index >= 0 && index < sushiList.length) {
       sushiList.splice(index, 1);
 
-      // ファイルを書き換え
       const updatedContent = `export const sushiList = ${JSON.stringify(sushiList, null, 2)};`;
       fs.writeFileSync(sushiListPath, updatedContent, 'utf8');
