+++ sushi-front/src/utils/sushiList.ts
@@ -1,17 +1,32 @@
 export const sushiList = [
   {
-    japanese: 'あんぱっちょぁんみょんじぇんっちゃん',
-    hiragana: 'あんぱっちょぁんみょんじぇんっちゃん',
-    romaji: [['a'], ['nnpa', 'npa', 'xnpa'], ['cchilyo', 'cchixyo', 'ccho', 'ccyo', 'ltuchilyo', 'ltuchixyo', 'ltucho', 'ltucyo', 'ltutilyo', 'ltutixyo', 'ltutyo', 'ttilyo', 'ttixyo', 'ttyo', 'xtuchilyo', 'xtuchixyo', 'xtucho', 'xtucyo', 'xtutilyo', 'xtutixyo', 'xtutyo'], ['xa', 'la'], ['nmilyo', 'nmixyo', 'nmyo', 'nnmilyo', 'nnmixyo', 'nnmyo', 'xnmilyo', 'xnmixyo', 'xnmyo'], ['nje', 'njile', 'njixe', 'njye', 'nnje', 'nnjile', 'nnjixe', 'nnjye', 'nnzile', 'nnzixe', 'nnzye', 'nzile', 'nzixe', 'nzye', 'xnje', 'xnjile', 'xnjixe', 'xnjye', 'xnzile', 'xnzixe', 'xnzye'], ['nccha', 'ncchila', 'ncchixa', 'nccya', 'nltucha', 'nltuchila', 'nltuchixa', 'nltucya', 'nltutila', 'nltutixa', 'nltutya', 'nnccha', 'nncchila', 'nncchixa', 'nnccya', 'nnltucha', 'nnltuchila', 'nnltuchixa', 'nnltucya', 'nnltutila', 'nnltutixa', 'nnltutya', 'nnttila', 'nnttixa', 'nnttya', 'nnxtucha', 'nnxtuchila', 'nnxtuchixa', 'nnxtucya', 'nnxtutila', 'nnxtutixa', 'nnxtutya', 'nttila', 'nttixa', 'nttya', 'nxtucha', 'nxtuchila', 'nxtuchixa', 'nxtucya', 'nxtutila', 'nxtutixa', 'nxtutya', 'xnccha', 'xncchila', 'xncchixa', 'xnccya', 'xnltucha', 'xnltuchila', 'xnltuchixa', 'xnltucya', 'xnltutila', 'xnltutixa', 'xnltutya', 'xnttila', 'xnttixa', 'xnttya', 'xnxtucha', 'xnxtuchila', 'xnxtuchixa', 'xnxtucya', 'xnxtutila', 'xnxtutixa', 'xnxtutya'], ['nn', 'xn']]
-  },
-  {
-    japanese: '西田',
-    hiragana: 'たなか',
-    romaji: [['ta'], ['na'], ['ka', 'ca']]
-  },
-  {
-    japanese: 'クソガキ',
-    hiragana: 'くそがき',
-    romaji: [['ku', 'cu', 'qu'], ['so'], ['ga'], ['ki']]
-  },
+    "japanese": "岩澤蓮司",
+    "hiragana": "いわさわれんじ",
+    "romaji": [
+      [
+        "i",
+        "yi"
+      ],
+      [
+        "wa"
+      ],
+      [
+        "sa"
+      ],
+      [
+        "wa"
+      ],
+      [
+        "re"
+      ],
+      [
+        "nji",
+        "nnji",
+        "nnzi",
+        "nzi",
+        "xnji",
+        "xnzi"
+      ]
+    ]
+  }
 ];
